Move index route to app folder (#149)
* move home to app folder

* ssr fix with countdown component

* ssr fix with category slider component

* cleanup
