+++ src/components/home/DealOfTheWeek.tsx
@@ -6,12 +6,12 @@ import Container from '~/layouts/Container';
 
 const { promotions } = data;
 
-const Countdown = dynamic(() => import('./Countdown'), { ssr: false });
-
 type DealOfTheWeekProps = {
   promotion: Promotion;
   imagePromotion?: string;
 };
+
+const Countdown = dynamic(() => import('./Countdown'), { ssr: false });
 
 const DealOfTheWeek = ({ promotion, imagePromotion }: DealOfTheWeekProps) => {
   const name = promotion.name;
@@ -24,8 +24,8 @@ const DealOfTheWeek = ({ promotion, imagePromotion }: DealOfTheWeekProps) => {
           <Image
             src={imagePromotion ? imagePromotion : promotions[0]?.image.src}
             alt={promotion.name}
-            layout="fill"
-            objectFit="cover"
+            fill
+            style={{ objectFit: 'cover' }}
           />
         </div>
         <Container>
